backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20
python-dotenv==1.0.0

# LangChain & LangGraph - Latest stable versions
langchain==0.3.13
langchain-google-genai==2.0.8
langgraph==0.2.60
google-generativeai==0.8.3

# Additional utilities
textblob==0.18.0.post0
